marketdata/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
redis>=5.0.0
firebase-admin>=6.4.0
google-cloud-secret-manager>=2.18.0
google-cloud-firestore>=2.14.0

